assigned permission to roles and perform authentication

+++ routes/web.php
@@ -41,7 +41,6 @@ Route::prefix('/Admin')->name('Admin.')->group(function () {
         // RoleAssignedtouser EndPoints
         Route::get('/RolesGiven',[RoleGivenController::class,'index'])->name('RolesGiven');
 
-        
         // users endpoint
         Route::get('/users',[UserController::class,'index'])->name('User');
         Route::get('/users/add',[UserController::class,'add'])->name('User.add');
@@ -84,19 +83,15 @@ Route::prefix('/Admin')->name('Admin.')->group(function () {
         Route::post('/AssignItem/edit-save',[ItemsController::class,'update'])->name('AssignItem.edit-save');
         Route::delete('/AssignItem/delete/{id}',[ItemsController::class,'destroy'])->name('AssignItem.delete');
 
-        // Route::middleware([AssignRole::class])->group(function(){
-        // Post endpoints
-        Route::get('/Posts/add', [PostsController::class,'add'])->name('Posts.add');
-        // });
-
-        Route::get('/Posts/edit/{id}',[PostsController::class,'edit'])->name('Posts.edit');
-        Route::delete('/Posts/delete/{id}',[PostsController::class,'destroy'])->name('Posts.delete');
-
+        Route::get('/Posts/add', [PostsController::class,'add'])->name('Posts.add')->middleware('CheckPermission:Post,Add');
+        Route::get('/Posts/edit/{id}',[PostsController::class,'edit'])->name('Posts.edit')->middleware('CheckPermission:Post,Edit');
+        Route::delete('/Posts/delete/{id}',[PostsController::class,'destroy'])->name('Posts.delete')->middleware('CheckPermission:Post,Delete');
 
         // comment endpoints
         Route::post('Comments/save',[CommentsController::class,'store'])->name('Comments.save');
-        Route::get('/Comments/edit/{id}',[CommentsController::class,'edit'])->name('Comments.edit');
-        Route::delete('/Comments/delete/{id}',[CommentsController::class,'destroy'])->name('Comments.delete');
+        Route::get('/Comments/edit/{id}',[CommentsController::class,'edit'])->name('Comments.edit')->middleware('CheckPermission:Comment,Edit');
+        Route::delete('/Comments/delete/{id}',[CommentsController::class,'destroy'])->name('Comments.delete')
+                        ->middleware('CheckPermission:Comment,Delete');
         
               
     });
